@@ -9,46 +9,50 @@
 		"preview": "vite preview",
 		"check": "svelte-check",
 		"check:watch": "svelte-check --watch",
-		"lint": "prettier --check --plugin=prettier-plugin-svelte . && eslint --ignore-path .gitignore .",
+		"lint": "prettier --check --plugin=prettier-plugin-svelte . && eslint .",
 		"lint:scss": "stylelint \"src/**/*.{css,scss,svelte}\"",
 		"prettier:check": "prettier --check --plugin=prettier-plugin-svelte .",
 		"format": "prettier --write --plugin=prettier-plugin-svelte .",
 		"prepare": "husky install"
 	},
 	"devDependencies": {
-		"@commitlint/config-conventional": "^18.4.2",
-		"@fontsource/lato": "^5.0.17",
-		"@fontsource/lora": "^5.0.8",
-		"@fontsource/source-sans-pro": "^5.0.8",
-		"@iconify/svelte": "^3.1.4",
-		"@rodneylab/sveltekit-components": "^3.0.0",
-		"@sveltejs/adapter-static": "^2.0.3",
-		"@sveltejs/kit": "^1.27.6",
+		"@commitlint/config-conventional": "^19.6.0",
+		"@fontsource/lato": "^5.1.0",
+		"@fontsource/lora": "^5.1.0",
+		"@fontsource/source-sans-pro": "^5.1.0",
+		"@iconify/svelte": "^4.0.2",
+		"@rodneylab/parsedown": "^1.4.3",
+		"@rodneylab/sveltekit-components": "^3.0.1",
+		"@sveltejs/adapter-static": "^3.0.6",
+		"@sveltejs/kit": "^2.9.0",
+		"@sveltejs/vite-plugin-svelte": "^5.0.1",
 		"@types/object-hash": "^3.0.6",
-		"commitlint": "^18.4.2",
-		"dayjs": "^1.11.10",
-		"eslint": "^8.54.0",
-		"eslint-config-prettier": "^9.0.0",
-		"eslint-plugin-svelte": "^2.35.0",
-		"husky": "^8.0.3",
-		"mdsvex": "^0.11.0",
+		"commitlint": "^19.6.0",
+		"dayjs": "^1.11.13",
+		"dompurify": "^3.2.2",
+		"eslint-config-prettier": "^9.1.0",
+		"eslint-plugin-svelte": "^2.46.1",
+		"globals": "^15.13.0",
+		"husky": "^9.1.7",
+		"jsdom": "^25.0.1",
 		"object-hash": "^3.0.0",
-		"postcss": "^8.4.31",
-		"postcss-html": "^1.5.0",
-		"prettier": "~3.1.0",
-		"prettier-plugin-svelte": "^3.1.0",
-		"reading-time": "~1.4.0",
-		"sass": "^1.69.5",
-		"stylelint": "^15.11.0",
-		"stylelint-config-recommended-scss": "^13.1.0",
-		"stylelint-config-standard": "^34.0.0",
-		"stylelint-scss": "^5.3.1",
-		"svelte": "^4.2.7",
-		"svelte-check": "^3.6.0",
-		"svelte-preprocess": "^5.1.0",
-		"vanilla-lazyload": "^17.8.5",
-		"vite": "^4.5.0",
-		"vite-imagetools": "^6.2.4"
+		"postcss": "^8.4.49",
+		"postcss-html": "^1.7.0",
+		"prettier": "~3.4.2",
+		"prettier-plugin-svelte": "^3.3.2",
+		"sass-embedded": "^1.82.0",
+		"stylelint": "^16.11.0",
+		"stylelint-config-recommended-scss": "^14.1.0",
+		"stylelint-config-standard": "^36.0.1",
+		"stylelint-scss": "^6.10.0",
+		"svelte": "^5.5.3",
+		"svelte-check": "^4.1.1",
+		"svelte-preprocess": "^6.0.3",
+		"typescript-eslint": "^8.17.0",
+		"vanilla-lazyload": "^19.1.3",
+		"vite": "^6.0.2",
+		"vite-imagetools": "^7.0.5",
+		"yaml": "^2.6.1"
 	},
 	"type": "module",
 	"repository": {
