@@ -1,10 +1,12 @@
 <script>
 	import Icon, { addCollection } from '@iconify/svelte/dist/OfflineIcon.svelte';
 
-	export let label = 'Telegram app icon';
-	export let colour = 'inherit';
-	export let ariaHidden = false;
-	export let width = 24;
+	let {
+		colour = 'inherit',
+		label = 'Telegram app icon',
+		ariaHidden = false,
+		width = 24,
+	} = $props();
 
 	addCollection({
 		prefix: 'simple-icons',
