@@ -18,12 +18,14 @@
 
 	import Footer from '$lib/components/Footer.svelte';
 	import Header from '$lib/components/Header.svelte';
+
+	let { children } = $props();
 </script>
 
 <div class="container">
 	<Header />
 	<main class="main-container">
-		<slot />
+		{@render children?.()}
 	</main>
 	<Footer />
 </div>
@@ -34,20 +36,20 @@
 		flex-wrap: wrap;
 		flex-direction: column;
 		justify-content: space-between;
-		padding: $spacing-4 $spacing-0 $spacing-0;
-		margin: $spacing-0 auto;
+		padding: variables.$spacing-4 variables.$spacing-0 variables.$spacing-0;
+		margin: variables.$spacing-0 auto;
 		min-height: 100vh;
 	}
 
 	.main-container {
-		max-width: $max-width-wrapper;
+		max-width: variables.$max-width-wrapper;
 		width: 100%;
-		margin: $spacing-0 auto;
+		margin: variables.$spacing-0 auto;
 	}
 
-	@media screen and (max-width: $desktop-breakpoint) {
+	@media screen and (max-width: variables.$desktop-breakpoint) {
 		.container {
-			margin: $spacing-0 $spacing-4;
+			margin: variables.$spacing-0 variables.$spacing-4;
 		}
 	}
 </style>
