@@ -1,9 +1,5 @@
-import { compile } from 'mdsvex';
-import { join, resolve } from 'node:path';
 export const BLOG_PATH = 'src/content/blog';
-
-const __dirname = resolve();
-const location = join(__dirname, BLOG_PATH);
+import { parse } from 'yaml';
 
 /**
  * Returns array of post slugs
@@ -46,26 +42,22 @@ export async function getPostsContent() {
 	}
 }
 
-/**
- * Returns an array of post metadata, with optional post body too.  Array is sort in reverse
- * chrononological order
- * @param {{slug: string; content: string;}[]} postsContent -
- * @param {boolean} body - if true the HTML post body is returned as well as meta
- */
-export const getPosts = async (postsContent, body = false) => {
-	const postPromises = postsContent.map(async (element) => {
-		const { content, slug } = element;
-		const transformedContent = await compile(content);
-		const { datePublished, lastUpdated, postTitle, seoMetaDescription } =
-			/** @type {{datePublished: string; lastUpdated: string; postTitle: string; seoMetaDescription: string;}} */ (
-				transformedContent.data.fm
-			);
-		let resultElement = { datePublished, lastUpdated, postTitle, seoMetaDescription, slug };
-		if (body) {
-			resultElement = { ...resultElement, body: transformedContent.code };
-		}
-		return resultElement;
-	});
-	const result = await Promise.all(postPromises);
-	return result.sort((a, b) => Date.parse(b.datePublished) - Date.parse(a.datePublished));
+export const separateFrontmatter = (markdown) => {
+	const frontmatterStartIndex = markdown.indexOf('---') + 3;
+	const frontmatterEndIndex = markdown.indexOf('---', frontmatterStartIndex);
+
+	if (frontmatterStartIndex !== -1 && frontmatterEndIndex !== -1) {
+		const parsedFrontmatter = parse(markdown.slice(frontmatterStartIndex, frontmatterEndIndex));
+
+		return {
+			frontmatter: parsedFrontmatter,
+			markdownBody: markdown.slice(frontmatterEndIndex + 3).trim(),
+		};
+	}
+
+	// assume there is no frontmatter
+	return {
+		frontmatter: null,
+		markdownBody: markdown,
+	};
 };
