@@ -92,13 +92,13 @@
 
 			li {
 				display: flex;
-				padding-left: $spacing-2;
-				font-size: $font-size-2;
+				padding-left: variables.$spacing-2;
+				font-size: variables.$font-size-2;
 			}
 		}
 	}
 
 	.contact-address {
-		margin-left: $spacing-2;
+		margin-left: variables.$spacing-2;
 	}
 </style>
