@@ -1,20 +1,10 @@
 /** @type {import('@sveltejs/kit').Config} */
 import adapter from '@sveltejs/adapter-static';
-import { mdsvex } from 'mdsvex';
-import preprocess from 'svelte-preprocess';
+import { vitePreprocess } from '@sveltejs/vite-plugin-svelte';
 
 const config = {
 	extensions: ['.svelte', '.md', '.svelte.md'],
-	preprocess: [
-		mdsvex({
-			extensions: ['.svelte.md', '.md', '.svx'],
-		}),
-		preprocess({
-			scss: {
-				prependData: "@import 'src/lib/styles/variables.scss';",
-			},
-		}),
-	],
+	preprocess: [vitePreprocess({})],
 	kit: {
 		adapter: adapter({ precompress: true }),
 	},
