@@ -1,12 +1,19 @@
 import { sveltekit } from '@sveltejs/kit/vite';
 import { imagetools } from 'vite-imagetools';
+import { defineConfig } from 'vite';
 
 /** @type {import('vite').UserConfig} */
-const config = {
+export default defineConfig({
+	css: {
+		preprocessorOptions: {
+			scss: {
+				additionalData: "@use 'variables.scss';",
+				loadPaths: ['src/lib/styles'],
+			},
+		},
+	},
 	define: {
 		'process.env.VITE_BUILD_TIME': JSON.stringify(new Date().toISOString()),
 	},
 	plugins: [sveltekit(), imagetools()],
-};
-
-export default config;
+});
